# Optimize `r_dyn_*_push_back()` for vector types