@DavisVaughan 2023-01-19
`-O2` and on an Intel mac